# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zyad Cloud API
  version: 0.1.0
paths:
...
  /api/v1/admin/landing/menus:
    get:
      summary: List landing menus
      operationId: listLandingMenus
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Menus retrieved successfully
    post:
      summary: Create landing menu
      operationId: createLandingMenu
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Menu created successfully
  /api/v1/admin/landing/menus/{id}:
    patch:
      summary: Update landing menu
      operationId: updateLandingMenu
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Menu updated successfully
    delete:
      summary: Delete landing menu
      operationId: deleteLandingMenu
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Menu deleted successfully
  /api/v1/admin/landing/menus/{id}/items:
    get:
      summary: List landing menu items
      operationId: listLandingMenuItems
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Menu items retrieved successfully
    post:
      summary: Create landing menu item
      operationId: createLandingMenuItem
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Menu item created successfully
  /api/v1/admin/landing/menus/{id}/items/reorder:
    put:
      summary: Reorder landing menu items
      operationId: reorderLandingMenuItems
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Menu items reordered successfully
  /api/v1/admin/landing/menus/{id}/items/{itemId}:
    patch:
      summary: Update landing menu item
      operationId: updateLandingMenuItem
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Menu item updated successfully
    delete:
      summary: Delete landing menu item
      operationId: deleteLandingMenuItem
      tags:
        - Landing Navigation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Menu item deleted successfully
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...